% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_slopeone.R
\name{build_slopeone}
\alias{build_slopeone}
\title{Build slope one model}
\usage{
build_slopeone(ratings, ...)
}
\arguments{
\item{ratings}{\code{data.table} of ratings. Should contain 3 columns: user_id (id of user, character), item_id (id of item, character) and rating (rating of item by user, integer or numeric)}
}
\value{
a \code{data.table} of (item_id1, item_id2, b, support) where b represents the average rating difference of 'item 2 rating' - 'item 1 rating'. support represents number of ratings used to compute b
}
\description{
Build slope one model
}
\examples{
set.seed(1)
in_train <- rep(TRUE, nrow(MovieLense))
in_train[sample(1:nrow(MovieLense), size = length(unique(MovieLense$user_id)) * 5)] <- FALSE

MovieLense_train <- MovieLense[(in_train)]

ratings <- normalize_ratings(MovieLense_train)
model <- build_slopeone(ratings$ratings)

}

