% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_ratings.R
\name{normalize_ratings}
\alias{normalize_ratings}
\title{Normalize rating table}
\usage{
normalize_ratings(ratings, ...)
}
\arguments{
\item{ratings}{\code{data.table} of ratings. Should contain 3 columns: user_id (id of user, character), item_id (id of item, character) and rating (rating of item by user, integer or numeric)}
}
\value{
List of 4 objects:
\itemize{
  \item global - global rating mean
  \item user - \code{data.table} of users' means (id, mean)
  \item item - \code{data.table} of items' means (id, mean)
  \item ratings - \code{data.table} of normalized ratings
}
}
\description{
Normalize rating table
}
\details{
Factor rating table into global mean + user mean + item mean. This is usually a preliminary step before building a model
}
\examples{
set.seed(1)
in_train <- rep(TRUE, nrow(MovieLense))
in_train[sample(1:nrow(MovieLense), size = length(unique(MovieLense$user_id)) * 5)] <- FALSE

MovieLense_train <- MovieLense[(in_train)]

ratings <- normalize_ratings(MovieLense_train)

}

