% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_slopeone.R
\name{predict_slopeone}
\alias{predict_slopeone}
\title{Predict ratings for multiple users and items given known ratings}
\usage{
predict_slopeone(model, targets, ratings, ...)
}
\arguments{
\item{model}{\code{data.table} of produced by \code{\link{build_slopeone}}}

\item{ratings}{\code{data.table} of known ratings. Should contain 3 columns: user_id (id of user, character), item_id (id of item, character) and rating (rating of item by user, integer or numeric)}

\item{target}{\code{data.table} of (user_id, item_id) to predict ratings}
}
\value{
a \code{data.table} containig (user_id, item_id, predicted_rating)
}
\description{
Predict ratings for multiple users and items given known ratings
}
\examples{
set.seed(1)
in_train <- rep(TRUE, nrow(MovieLense))
in_train[sample(1:nrow(MovieLense), size = length(unique(MovieLense$user_id)) * 5)] <- FALSE

MovieLense_train <- MovieLense[(in_train)]

ratings <- normalize_ratings(MovieLense_train)
model <- build_slopeone(ratings$ratings)

predictions <- predict_slopeone(model, MovieLense[(!in_train), c(1, 2), with = FALSE], ratings$ratings)

}

