% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_slopeone_for_user.R
\name{predict_slopeone_for_user}
\alias{predict_slopeone_for_user}
\title{Predict score for target_item_id given the known ratings of a single user}
\usage{
predict_slopeone_for_user(model, target_item_id, ratings)
}
\arguments{
\item{model}{\code{data.table} of produced by \code{\link{build_slopeone}}}

\item{target_item_id}{target item id to predict rating}

\item{ratings}{\code{data.table} of user's known ratings. Should contain 3 columns: user_id (id of user, character), item_id (id of item, character) and rating (rating of item by user, integer or numeric)}
}
\value{
predicted rating score
}
\description{
Predict score for target_item_id given the known ratings of a single user
}
\examples{
set.seed(1)
in_train <- rep(TRUE, nrow(MovieLense))
in_train[sample(1:nrow(MovieLense), size = length(unique(MovieLense$user_id)) * 5)] <- FALSE

MovieLense_train <- MovieLense[(in_train)]

ratings <- normalize_ratings(MovieLense_train)
model <- build_slopeone(ratings$ratings)
predict_slopeone_for_user(model, "Volcano (1997)", ratings$ratings[user_id == "1", c(2, 3), with = FALSE])

}

