% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnormalize_ratings.R
\name{unnormalize_ratings}
\alias{unnormalize_ratings}
\title{Un-normalize rating table}
\usage{
unnormalize_ratings(normalized, ratings)
}
\arguments{
\item{normalized}{normalization information generated by \code{\link{normalize_ratings}}}

\item{ratings}{\code{data.table} of ratings. Should contain 3 columns: userI_id (id of user, character), item_id (id of item, character) and predicted_rating (predicted rating of item by user, integer or numeric)}
}
\value{
a ratings \code{data.table} after un-normalization
}
\description{
Un-normalize rating table
}
\details{
Need to apply this after prediction is made using a model built from normalized ratings
}
\examples{
set.seed(1)
in_train <- rep(TRUE, nrow(MovieLense))
in_train[sample(1:nrow(MovieLense), size = length(unique(MovieLense$user_id)) * 5)] <- FALSE

MovieLense_train <- MovieLense[(in_train)]

ratings <- normalize_ratings(MovieLense_train)
model <- build_slopeone(ratings$ratings)

predictions <- predict_slopeone(model, MovieLense[(!in_train), c(1, 2), with = FALSE], ratings$ratings)
real_ratings <- unnormalize_ratings(normalized = ratings, ratings = predictions)

rmse <- sqrt(mean((real_ratings$predicted_rating - MovieLense[(!in_train)]$rating) ^ 2))

}

